Refactor callback middleware import and create Callback class for improved functionality
